+++ functions/src/callable/index.ts
@@ -1,2 +1,1 @@
 export { contactus } from './emails/support';
-export { messaging } from './messagings/messaging';