+++ functions/src/index.ts
@@ -1,8 +1,5 @@
 import * as admin from 'firebase-admin';
 
-admin.initializeApp({
-  credential: admin.credential.applicationDefault(),
-  databaseURL: 'https://lightcutoff.firebaseio.com'
-});
+admin.initializeApp();
 
 export * from './callable';